Make method for updating scanned docs non-static
